+++ TerraLeague/UI/TeleportUI.cs
@@ -503,7 +503,7 @@ namespace TerraLeague.UI
 
                 if (hasTP)
                 {
-                    modPlayer.sumCooldowns[i] = modPlayer.sumSpells[i].GetCooldown() * 60;
+                    modPlayer.sumCooldowns[i] = (int)(modPlayer.sumSpells[i].GetCooldown() * 60);
                 }
             }
         }
